Move error module to separate file
