+++ src/assets.rs
